crypto_investment_assistant/requirements.txt
requests==2.28.2
pandas==1.5.3
schedule==1.1.0
python-telegram-bot==20.3
scikit-learn==1.2.2
numpy==1.23.5